requirements.txt
Flask==3.0.0
newspaper3k==0.2.8
transformers==4.35.2
torch==2.1.1
textblob==0.17.1
sentencepiece==0.1.99
